make it suck less
